+++ ci/appveyor_run_tests.ps1
@@ -44,6 +44,23 @@ if ($CS_STATUS -ne 0) {
     Write-Host "Embedded tests failed" -ForegroundColor "Red"
 }
 
+if ($env:BUILD_OPTS -eq "--xplat"){
+    if ($env:PLATFORM -eq "x64") {
+         $DOTNET_CMD = "dotnet"
+    }
+    else{
+         $DOTNET_CMD = "c:\Program Files (x86)\dotnet\dotnet"
+    }
+
+    # Run Embedded tests for netcoreapp2.0 (OpenCover currently does not supports dotnet core)
+    Write-Host ("Starting embedded tests for netcoreapp2.0") -ForegroundColor "Green"
+    &$DOTNET_CMD .\src\embed_tests\bin\netcoreapp2.0_publish\Python.EmbeddingTest.dll
+    $CS_STATUS = $LastExitCode
+    if ($CS_STATUS -ne 0) {
+        Write-Host "Embedded tests for netcoreapp2.0 failed" -ForegroundColor "Red"
+    }
+}
+
 # Set exit code to fail if either Python or Embedded tests failed
 if ($PYTHON_STATUS -ne 0 -or $CS_STATUS -ne 0) {
     Write-Host "Tests failed" -ForegroundColor "Red"