+++ src/embed_tests/Program.cs
@@ -1,0 +1,19 @@
+using System;
+
+using NUnit.Common;
+
+using NUnitLite;
+
+namespace Python.EmbeddingTest
+{
+    public class Program
+    {
+        public static int Main(string[] args)
+        {
+            return new AutoRun(typeof(Program).Assembly).Execute(
+                args,
+                new ExtendedTextWrapper(Console.Out),
+                Console.In);
+        }
+    }
+}