+++ src/runtime/runtime.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Runtime.InteropServices;
 using System.Security;
 using System.Text;
@@ -9,6 +9,27 @@ namespace Python.Runtime
     internal static class NativeMethods
     {
 #if MONO_LINUX || MONO_OSX
+#if NETSTANDARD
+        private static int RTLD_NOW = 0x2;
+#if MONO_LINUX
+        private static int RTLD_GLOBAL = 0x100;
+        private static IntPtr RTLD_DEFAULT = IntPtr.Zero;
+        private const string NativeDll = "libdl.so";
+        public static IntPtr LoadLibrary(string fileName)
+        {
+            return dlopen($"lib{fileName}.so", RTLD_NOW | RTLD_GLOBAL);
+        }
+#elif MONO_OSX
+        private static int RTLD_GLOBAL = 0x8;
+        private const string NativeDll = "/usr/lib/libSystem.dylib"
+        private static IntPtr RTLD_DEFAULT = new IntPtr(-2);
+
+        public static IntPtr LoadLibrary(string fileName)
+        {
+            return dlopen($"lib{fileName}.dylib", RTLD_NOW | RTLD_GLOBAL);
+        }
+#endif
+#else
         private static int RTLD_NOW = 0x2;
         private static int RTLD_SHARED = 0x20;
 #if MONO_OSX
@@ -23,6 +44,8 @@ namespace Python.Runtime
         {
             return dlopen(fileName, RTLD_NOW | RTLD_SHARED);
         }
+#endif
+
 
         public static void FreeLibrary(IntPtr handle)
         {
@@ -48,16 +71,16 @@ namespace Python.Runtime
             return res;
         }
 
-        [DllImport(NativeDll)]
-        private static extern IntPtr dlopen(String fileName, int flags);
+        [DllImport(NativeDll, CallingConvention = CallingConvention.Cdecl, CharSet = CharSet.Ansi)]
+        public static extern IntPtr dlopen(String fileName, int flags);
 
-        [DllImport(NativeDll)]
+        [DllImport(NativeDll, CallingConvention = CallingConvention.Cdecl, CharSet = CharSet.Ansi)]
         private static extern IntPtr dlsym(IntPtr handle, String symbol);
 
-        [DllImport(NativeDll)]
+        [DllImport(NativeDll, CallingConvention = CallingConvention.Cdecl)]
         private static extern int dlclose(IntPtr handle);
 
-        [DllImport(NativeDll)]
+        [DllImport(NativeDll, CallingConvention = CallingConvention.Cdecl)]
         private static extern IntPtr dlerror();
 #else // Windows
         private const string NativeDll = "kernel32.dll";
@@ -158,7 +181,7 @@ namespace Python.Runtime
 
         public static readonly string PythonDLL = _PythonDll;
 
-#if PYTHON_WITHOUT_ENABLE_SHARED
+#if PYTHON_WITHOUT_ENABLE_SHARED && !NETSTANDARD
         internal const string _PythonDll = "__Internal";
 #else
         internal const string _PythonDll = dllBase + dllWithPyDebug + dllWithPyMalloc;
@@ -298,11 +321,13 @@ namespace Python.Runtime
             Error = new IntPtr(-1);
 
             IntPtr dllLocal = IntPtr.Zero;
+
             if (_PythonDll != "__Internal")
             {
                 dllLocal = NativeMethods.LoadLibrary(_PythonDll);
             }
             _PyObject_NextNotImplemented = NativeMethods.GetProcAddress(dllLocal, "_PyObject_NextNotImplemented");
+
 #if !(MONO_LINUX || MONO_OSX)
             if (dllLocal != IntPtr.Zero)
             {
@@ -497,7 +522,7 @@ namespace Python.Runtime
         /// </summary>
         internal static unsafe void XIncref(IntPtr op)
         {
-#if PYTHON_WITH_PYDEBUG
+#if PYTHON_WITH_PYDEBUG || NETSTANDARD
             Py_IncRef(op);
             return;
 #else
@@ -518,7 +543,7 @@ namespace Python.Runtime
 
         internal static unsafe void XDecref(IntPtr op)
         {
-#if PYTHON_WITH_PYDEBUG
+#if PYTHON_WITH_PYDEBUG || NETSTANDARD
             Py_DecRef(op);
             return;
 #else